+++ src/AppForSEII2526.API/DTOs/DTOsPedirBocadillo/ArticuloPedidoDTO.cs
@@ -8,14 +8,14 @@ namespace AppForSEII2526.API.DTOs.DTOs_PedirBocadillo
 
         [Required]
         public int Cantidad { get; set; }
-        public decimal PVP { get; set; }
+        public float PVP { get; set; }
         public string TipoPan { get; set; }
 
         public ArticuloPedidoDTO()
         {
         }
 
-        public ArticuloPedidoDTO(int id, string nombreBocadillo, int cantidad, decimal pVP, string tipoPan)
+        public ArticuloPedidoDTO(int id, string nombreBocadillo, int cantidad, float pVP, string tipoPan)
         {
             Id = id;
             this.nombreBocadillo = nombreBocadillo;
